Merge remote-tracking branch 'origin/caf/caf-wlan/master' into jb_mr1

* origin/caf/caf-wlan/master:
  Wlan: Release 3.0.10.061
  wlan: trim the meta-information from logging packets
  wlan: HDD API should consult ini parameters to check logging status
  prima: ioctl enable or disable 20/40 bss coex IE in TDLS setup Req/Rsp
  prima: Teardown TDLS links in case of concurrency

CRs-Fixed: 688141

+++ CORE/SVC/inc/wlan_logging_sock_svc.h
@@ -45,6 +45,6 @@ int wlan_logging_sock_activate_svc(int log_fe_to_console, int num_buf);
 int wlan_logging_flush_pkt_queue(void);
 int wlan_logging_sock_deactivate_svc(void);
 int wlan_log_to_user(VOS_TRACE_LEVEL log_level, char *to_be_sent, int length);
-int wlan_queue_logpkt_for_app(vos_pkt_t *pPacket, uint8 pkt_type);
+int wlan_queue_logpkt_for_app(vos_pkt_t *pPacket, uint32 pkt_type);
 
 #endif /* WLAN_LOGGING_SOCK_SVC_H */
